faceblur-auto/requirements.txt
cmake
gradio>=4.29.0
face_recognition
opencv-python
numpy
Pillow
